+++ weave/legacy/panel.py
@@ -4,9 +4,9 @@ import typing
 from tarfile import DEFAULT_FORMAT
 
 from weave.legacy import api as weave
-from weave import storage, weave_internal
+from weave import weave_internal
 from weave import weave_types as types
-from weave.legacy import graph, panel_util, errors
+from weave.legacy import graph, panel_util, errors, storage
 
 
 def run_variable_lambdas(
