+++ weave/legacy/wandb_interface/wandb_stream_table.py
@@ -15,8 +15,9 @@ from wandb.sdk.lib.ipython import _get_python_type
 from wandb.sdk.lib.paths import LogicalPath
 from wandb.sdk.lib.printer import get_printer
 
-from weave import storage, weave_types
+from weave import weave_types
 from weave.legacy import (
+    storage, 
     errors, 
     artifact_base,
     box,
